avcodec/dxa: use uint8_t for type of arrays

+++ libavcodec/dxa.c
@@ -45,8 +45,8 @@ typedef struct DxaDecContext {
     uint32_t pal[256];
 } DxaDecContext;
 
-static const int shift1[6] = { 0, 8, 8, 8, 4, 4 };
-static const int shift2[6] = { 0, 0, 8, 4, 0, 4 };
+static const uint8_t shift1[6] = { 0, 8, 8, 8, 4, 4 };
+static const uint8_t shift2[6] = { 0, 0, 8, 4, 0, 4 };
 
 static int decode_13(AVCodecContext *avctx, DxaDecContext *c, uint8_t* dst,
                      int stride, uint8_t *src, int srcsize, uint8_t *ref)
